feat: rename settings app by global_settings
